+++ Core/datadefs.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "datastructures.h"
 
-extern vector<vector<int>> segdefs;
+extern vector<vector<int64_t>> segdefs;
 extern vector<transdef> transdefs;
 extern unordered_map<string, uint16_t> nodenames;
 extern vector<vector<vector<int>>> palette_nodes;
